vsock: include the errno when C.bind_sockaddr_vm
All cgo functions can be called in a multiple assignment context to get both
the return value and the errno. Do so and log the errno as well on failure
since it will be more interesting than the return value (which is always -1 on
error, but continue logging it anyway...)

Wrap the long line in a way which gofmt seems happy with.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>