Do Rx incrementally in configurable sized batches
It is now possible to send/receive very large numbers of bytes over a single
connection, they will be written in checking given by the bufSize (-b) option.

This partially undoes "Use a single Reader for the entire client side"
(579e02187a50). This is deliberate since it removes the buffered Reader from
the main loop, ensuring that our attempts to read differing sized buffers are
not defeated by buffing within the buffered Reader.

It is now safe to construct the Reader solely for use in reading the BYE
message since nothing previous will be buffering anything.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>