hvsock: Add support for the new Hyper-V socket implementation
There are new patches for Hyper-V socket, which work on top of the
common infrastructure setup for VMware/virtio sockets.  This new
implementation use the same Address family as vsock, but we can't
directly re-use the vsock implementation as we still need to emulate
CloseRead()/CloseWrite(), which are not supported on older Windows
versions.

Instead, just provide alternative implementation for connect(),
listen(), and accept(). This is a bit of code duplication, which
hopefully will go away once we don't have to support Windows build
10586 anymore.

We detect which implementation to use by probing the for AF_HYPERV.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>