Simplify sock_stress
The stress test code now takes client and server specs in URL
for, e.g., vsock://1234:5678, and various variations. This
allows us to remove a bunch of code where we try to detect
which scheme to use.

Since the vsock and hvsock packages can now be used on all
platforms, there is less platform specific code needed for
sock_stress and the code can be simplified.

Abstract both vsock and hvsock into a Sock interface which,
later can also be implemented for TCP or other streams.

While at it also change the build to use the GOPATH instead
of listing all the individual source files.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>